Reworked Game UI and added additional features
A complete rework of the base game, changed UI from using JOptionPane to Gradle, rewrote the background math and adjusted most of the major features.
The main gameplay loop is still relatively uneventful, but as the game was mostly made to learn the basics of gamedev and UI in java and not as a passion project i won't be finishing it.